move models to right place
